removed unused codes, renamed component; sorted styled-components
