+++ client/src/components/FavoriteCoinsCard.js
