+++ client/src/components/FavoriteCoinsCard.md
