+++ client/src/components/Header.md
@@ -1,3 +1,3 @@
 ```jsx
-<ShowLogo />
+<Header />
 ```
